ml/requirements.txt
# Reinforcement Learning
gymnasium==0.29.1
stable-baselines3[extra]==2.1.0
torch==2.1.0

# Logging
tensorboard==2.15.1

# Data Handling
pandas==2.1.1

# Database
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
python-dotenv==1.0.0
